+++ lib/adf_info.py
@@ -78,6 +78,21 @@ class AdfInfo(AdfConfig):
         # Expand CAM climo info variable strings:
         self.expand_references(self.__cam_climo_info)
 
+        # Add CVDP info to object:
+        self.__cvdp_info = self.read_config_var("diag_cvdp_info")
+
+        # Expand CVDP climo info variable strings:
+        if self.__cvdp_info is not None:
+            self.expand_references(self.__cvdp_info)
+        # End if
+
+        # Add MDTF info to object:
+        self.__mdtf_info = self.read_config_var("diag_mdtf_info")
+
+        if self.__mdtf_info is not None:
+            self.expand_references(self.__mdtf_info)
+        # End if
+        
         # Check if inputs are of the correct type:
         # -------------------------------------------
 
@@ -232,49 +247,48 @@ class AdfInfo(AdfConfig):
 
             # Check if history file path exists:
             if any(baseline_hist_locs):
-                for hist_str in baseline_hist_str:
-                    starting_location = Path(baseline_hist_locs)
-                    file_list = sorted(starting_location.glob("*" + hist_str + ".*.nc"))
-                    # Partition string to find exactly where h-number is
-                    # This cuts the string before and after the `{hist_str}.` sub-string
-                    # so there will always be three parts:
-                    # before sub-string, sub-string, and after sub-string
-                    # Since the last part always includes the time range, grab that with last index (2)
-                    # NOTE: this is based off the current CAM file name structure in the form:
-                    #  $CASE.cam.h#.YYYY<other date info>.nc
+                hist_str = baseline_hist_str[0]
+                starting_location = Path(baseline_hist_locs)
+                file_list = sorted(starting_location.glob("*" + hist_str + ".*.nc"))
+                # Partition string to find exactly where h-number is
+                # This cuts the string before and after the `{hist_str}.` sub-string
+                # so there will always be three parts:
+                # before sub-string, sub-string, and after sub-string
+                # Since the last part always includes the time range, grab that with last index (2)
+                # NOTE: this is based off the current CAM file name structure in the form:
+                #  $CASE.cam.h#.YYYY<other date info>.nc
 
-                    base_climo_yrs = [
-                        int(str(i).partition(f"{hist_str}.")[2][0:4]) for i in file_list
-                    ]
-                    base_climo_yrs = sorted(np.unique(base_climo_yrs))
-                    base_found_syr = int(base_climo_yrs[0])
-                    base_found_eyr = int(base_climo_yrs[-1])
+                base_climo_yrs = [
+                    int(str(i).partition(f"{hist_str}.")[2][0:4]) for i in file_list
+                ]
+                base_climo_yrs = sorted(np.unique(base_climo_yrs))
+                base_found_syr = int(base_climo_yrs[0])
+                base_found_eyr = int(base_climo_yrs[-1])
 
-                    # Check if start or end year is missing. If so then just assume it is the
-                    # start or end of the entire available model data.
-                    if syear_baseline is None:
-                        msg = f"No given start year for {data_name}, "
-                        msg += f"using first found year: {base_found_syr}"
-                        print(msg)
-                        syear_baseline = base_found_syr
-                    if syear_baseline not in base_climo_yrs:
-                        msg = f"Given start year '{syear_baseline}' is not in current dataset "
-                        msg += f"{data_name}, using first found year: {base_climo_yrs[0]}\n"
-                        print(msg)
-                        syear_baseline = base_found_syr
+                # Check if start or end year is missing. If so then just assume it is the
+                # start or end of the entire available model data.
+                if syear_baseline is None:
+                    msg = f"No given start year for {data_name}, "
+                    msg += f"using first found year: {base_found_syr}"
+                    print(msg)
+                    syear_baseline = base_found_syr
+                if syear_baseline not in base_climo_yrs:
+                    msg = f"Given start year '{syear_baseline}' is not in current dataset "
+                    msg += f"{data_name}, using first found year: {base_climo_yrs[0]}\n"
+                    print(msg)
+                    syear_baseline = base_found_syr
 
-                    if eyear_baseline is None:
-                        msg = f"No given end year for {data_name}, "
-                        msg += f"using last found year: {base_found_eyr}"
-                        print(msg)
-                        eyear_baseline = base_found_eyr
-                    if eyear_baseline not in base_climo_yrs:
-                        msg = f"Given end year '{eyear_baseline}' is not in current dataset "
-                        msg += f"{data_name}, using last found year: {base_climo_yrs[-1]}\n"
-                        print(msg)
-                        eyear_baseline = base_found_eyr
+                if eyear_baseline is None:
+                    msg = f"No given end year for {data_name}, "
+                    msg += f"using last found year: {base_found_eyr}"
+                    print(msg)
+                    eyear_baseline = base_found_eyr
+                if eyear_baseline not in base_climo_yrs:
+                    msg = f"Given end year '{eyear_baseline}' is not in current dataset "
+                    msg += f"{data_name}, using last found year: {base_climo_yrs[-1]}\n"
+                    print(msg)
+                    eyear_baseline = base_found_eyr
 
-                # end for hist_str
             # End if any(baseline_hist_locs)
 
             # Grab baseline nickname
@@ -385,51 +399,50 @@ class AdfInfo(AdfConfig):
             # Check if history file path exists:
             hist_str_case = cam_hist_str[case_idx]
             if any(cam_hist_locs):
-                for hist_str in hist_str_case:
+                hist_str = hist_str_case[0]
 
-                    # Get climo years for verification or assignment if missing
-                    starting_location = Path(cam_hist_locs[case_idx])
-                    file_list = sorted(starting_location.glob("*" + hist_str + ".*.nc"))
-                    # Partition string to find exactly where h-number is
-                    # This cuts the string before and after the `{hist_str}.` sub-string
-                    # so there will always be three parts:
-                    # before sub-string, sub-string, and after sub-string
-                    # Since the last part always includes the time range, grab that with last index (2)
-                    # NOTE: this is based off the current CAM file name structure in the form:
-                    #  $CASE.cam.h#.YYYY<other date info>.nc
-                    case_climo_yrs = [
-                        int(str(i).partition(f"{hist_str}.")[2][0:4]) for i in file_list
-                    ]
-                    case_climo_yrs = sorted(np.unique(case_climo_yrs))
+                # Get climo years for verification or assignment if missing
+                starting_location = Path(cam_hist_locs[case_idx])
+                file_list = sorted(starting_location.glob("*" + hist_str + ".*.nc"))
+                # Partition string to find exactly where h-number is
+                # This cuts the string before and after the `{hist_str}.` sub-string
+                # so there will always be three parts:
+                # before sub-string, sub-string, and after sub-string
+                # Since the last part always includes the time range, grab that with last index (2)
+                # NOTE: this is based off the current CAM file name structure in the form:
+                #  $CASE.cam.h#.YYYY<other date info>.nc
+                case_climo_yrs = [
+                    int(str(i).partition(f"{hist_str}.")[2][0:4]) for i in file_list
+                ]
+                case_climo_yrs = sorted(np.unique(case_climo_yrs))
 
-                    case_found_syr = int(case_climo_yrs[0])
-                    case_found_eyr = int(case_climo_yrs[-1])
+                case_found_syr = int(case_climo_yrs[0])
+                case_found_eyr = int(case_climo_yrs[-1])
 
-                    # Check if start or end year is missing.  If so then just assume it is the
-                    # start or end of the entire available model data.
-                    if syear is None:
-                        msg = f"No given start year for {case_name}, "
-                        msg += f"using first found year: {case_found_syr}"
-                        print(msg)
-                        syear = case_found_syr
-                    if syear not in case_climo_yrs:
-                        msg = f"Given start year '{syear}' is not in current dataset "
-                        msg += f"{case_name}, using first found year: {case_climo_yrs[0]}\n"
-                        print(msg)
-                        syear = case_found_syr
-                    # End if
-                    if eyear is None:
-                        msg = f"No given end year for {case_name}, "
-                        msg += f"using last found year: {case_found_eyr}"
-                        print(msg)
-                        eyear = case_found_eyr
-                    if eyear not in case_climo_yrs:
-                        msg = f"Given end year '{eyear}' is not in current dataset "
-                        msg += f"{case_name}, using last found year: {case_climo_yrs[-1]}\n"
-                        print(msg)
-                        eyear = case_found_eyr
-                    # End if
-                # end for hist_str_file in enumerate(hist_str):
+                # Check if start or end year is missing.  If so then just assume it is the
+                # start or end of the entire available model data.
+                if syear is None:
+                    msg = f"No given start year for {case_name}, "
+                    msg += f"using first found year: {case_found_syr}"
+                    print(msg)
+                    syear = case_found_syr
+                if syear not in case_climo_yrs:
+                    msg = f"Given start year '{syear}' is not in current dataset "
+                    msg += f"{case_name}, using first found year: {case_climo_yrs[0]}\n"
+                    print(msg)
+                    syear = case_found_syr
+                # End if
+                if eyear is None:
+                    msg = f"No given end year for {case_name}, "
+                    msg += f"using last found year: {case_found_eyr}"
+                    print(msg)
+                    eyear = case_found_eyr
+                if eyear not in case_climo_yrs:
+                    msg = f"Given end year '{eyear}' is not in current dataset "
+                    msg += f"{case_name}, using last found year: {case_climo_yrs[-1]}\n"
+                    print(msg)
+                    eyear = case_found_eyr
+                # End if
             # End if
 
             # Update climo year lists in case anything changed
@@ -523,12 +536,15 @@ class AdfInfo(AdfConfig):
 
     #########
     def hist_str_to_list(self, conf_var, conf_val):
+        """
+        Make hist_str a nested list [ncases,nfiles] of the given value(s)
+        """
         if isinstance(self.get_cam_info("hist_str", required=True), list):
             hist_str = conf_val
         else:  # one case, one hist str
             hist_str = [
                 conf_val
-            ]  # make a nested list [ncases,nfiles] of the given value
+            ]  
         self.__cam_climo_info[conf_var] = [hist_str]
         # -----------------------------------------
 
@@ -625,8 +641,7 @@ class AdfInfo(AdfConfig):
     def get_basic_info(self, var_str, required=False):
         """
         Return the config variable from 'diag_basic_info' as requested by
-        the user.  This function assumes that if the user is requesting it,
-        then it must be required.
+        the user.  
         """
 
         return self.read_config_var(
@@ -639,9 +654,7 @@ class AdfInfo(AdfConfig):
     def get_cam_info(self, var_str, required=False):
         """
         Return the config variable from 'diag_cam_climo' as requested by
-        the user.  This function assumes that if the user is requesting it,
-        then it must be required. (Note: This statement contradicts the default value of required=False)
-        """
+        the user.  """
 
         return self.read_config_var(
             var_str, conf_dict=self.__cam_climo_info, required=required
@@ -692,6 +705,37 @@ class AdfInfo(AdfConfig):
 
     #########
 
+    # Utility function to access expanded 'diag_cvdp_info' variables
+    def get_cvdp_info(self, var_str, required=False):
+        """
+        Return the config variable from 'diag_cvdp_info' as requested by
+        the user. If 'diag_cvdp_info' is not found then try grabbing the
+        variable from the top level of the YAML config file dictionary
+        instead.
+        """
+
+        return self.read_config_var(
+            var_str, conf_dict=self.__cvdp_info, required=required
+        )
+
+    #########
+
+    # Utility function to access expanded 'diag_mdtf_info' variables
+    def get_mdtf_info(self, var_str, required=False):
+        """
+        Return the config variable from 'diag_mdtf_info' as requested by
+        the user. If 'diag_mdtf_info' is not found then try grabbing the
+        variable from the top level of the YAML config file dictionary
+        instead.
+        """
+
+        return self.read_config_var(
+            var_str, conf_dict=self.__mdtf_info, required=required
+        )
+ 
+        
+    #########
+   
     # Utility function to grab climo years from pre-made time series files:
     def get_climo_yrs_from_ts(self, input_ts_loc, case_name):
         """
@@ -719,6 +763,7 @@ class AdfInfo(AdfConfig):
 
         # Search for first variable in var_list to get a time series file to read
         # NOTE: it is assumed all the variables have the same dates!
+        # Also, it is assumed that only h0 files should be climo-ed.
         ts_files = sorted(input_location.glob(f"{case_name}*.{var_list[0]}.*nc"))
 
         # Read in file(s)
